+++ controllers/ladder.js
@@ -3,6 +3,7 @@ const { UNEXPECTED_ERROR, LADDER_ADDED, LADDER_UPDATED, LADDER_DELETED } = requi
 const { checkFile, makeResponse, listData, searchData, writeExcelFile } = require("../services/general");
 const LadderModel = require("../models/Ladder");
 const {unlinkSync} = require("fs");
+const { isObjectIdOrHexString } = require("mongoose");
 
 
 
@@ -75,7 +76,13 @@ module.exports.update = async (req, res) => {
         if(!givenObject.startingDate) return res.status(INVALID).send(makeResponse("Starting Date is Empty"));
         if(!givenObject.endingDate) return res.status(INVALID).send(makeResponse("Ending Date is Empty"));
 
-        let value = await LadderModel.updateOne({_id: id}, givenObject).catch((e) => {console.log(e)});
+        let value = await LadderModel.findOne({_id: id}).catch((e) => {console.log(e)});
+        if(value && value.picture){
+            try{
+                unlinkSync(value.picture);
+            }catch(e){}
+        }
+        value = await LadderModel.updateOne({_id: id}, givenObject).catch((e) => {console.log(e)});
         if(value.modifiedCount) return res.status(OK).send(makeResponse(LADDER_UPDATED));
         return res.status(NOT_FOUND).send(makeResponse(UNEXPECTED_ERROR));
     }
@@ -87,9 +94,14 @@ module.exports.delete = async (req, res) => {
 
     let id = req.params?.id || "";
     if(!id) return res.status(INVALID).send(makeResponse(UNEXPECTED_ERROR));
+    if(!isObjectIdOrHexString(id)) return res.status(INVALID).send(makeResponse("Invalid ID"));
     
     let value=await LadderModel.findOne({_id: id}).catch((e) => {console.log(e)});
-    if(value && value.picture){ unlinkSync(value.picture) };
+    if(value && value.picture){ 
+        try{
+            unlinkSync(value.picture) 
+        }catch(e){}
+    };
     
     value = await LadderModel.deleteOne({_id: id}).catch((e) => {console.log(e)});
     if(!value.deletedCount) return res.status(OK).send(makeResponse(LADDER_DELETED));
@@ -101,16 +113,18 @@ module.exports.show = async (req, res) => {
 
     let id = req.params?.id || "";
     if(!id) return res.status(INVALID).send(makeResponse(UNEXPECTED_ERROR));
+    if(!isObjectIdOrHexString(id)) return res.status(INVALID).send(makeResponse("Invalid ID"));
 
     let value = await LadderModel.findOne({_id: id}).catch((e) => {console.log(e)});
-    return res.status(OK).send(makeResponse("", value));
+    if(value) return res.status(OK).send(makeResponse("", value));
+    return res.status(NOT_FOUND).send(makeResponse("No data found"));
 }
 
 module.exports.list = async (req, res) => {
     if(!req.auth?.auth) return res.status(UNAUTHORIZED).send(makeResponse(AUTH_FAILED));
 
     let page_number = req.query?.pageNum || 1;
-    res.status(OK).send(makeResponse("", listData(LadderModel, page_number)));
+    res.status(OK).send(makeResponse("", await listData(LadderModel, page_number)));
 }
 
 
@@ -121,15 +135,15 @@ module.exports.list = async (req, res) => {
 module.exports.searchData = async (req, res) => {
     if(!req.auth?.auth) return res.status(UNAUTHORIZED).send(makeResponse(AUTH_FAILED));
 
-    let filter = req.body?.filter.toLowerCase() || "";
+    let filter = req.body?.filter?.toLowerCase() || "";
     let fields = ["name", "gameName"];
-    res.status(OK).send(makeResponse("", searchData(LadderModel, filter, fields)));
+    res.status(OK).send(makeResponse("", await searchData(LadderModel, filter, fields)));
 }
 
 module.exports.downloadExcel = async (req, res) => {
     if(!req.auth?.auth) return res.status(UNAUTHORIZED).send(makeResponse(AUTH_FAILED));
 
     let fields = ["name", "gameName", "entryFee", "prize", "teamSize", "totalTeams", "status", "startingDate", "endingDate"];
-    let value = LadderModel.find().catch((e)=>{console.log(e)});
-    return res.status(OK).send(makeResponse("File Written successfully", writeExcelFile(value, fields)));
+    let value = await LadderModel.find().catch((e)=>{console.log(e)});
+    return res.status(OK).send(makeResponse("File Written successfully", await writeExcelFile(value, fields)));
 }
